Rename test classes to reflect their new purpuse
