--- frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/classtag/ChildInterface.java
@@ -1,5 +1,0 @@
-package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.classtag;
-
-public interface ChildInterface extends InterfaceWithJavaDocTag {
-
-}
