--- frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/classtag/ClassWithoutJavaDocTag.java
@@ -1,4 +1,0 @@
-package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.classtag;
-
-public class ClassWithoutJavaDocTag {
-}
