--- frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/classtag/InheriterFromGrandparent.java
@@ -1,5 +1,0 @@
-package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.classtag;
-
-public class InheriterFromGrandparent extends InheriterFromParent {
-
-}
