--- frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/classtag/ParentOverrider.java
@@ -1,11 +1,0 @@
-package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.classtag;
-
-import nl.nn.adapterframework.doc.IbisDoc;
-
-/**
- * @author martijn
- *
- */
-@IbisDoc("ParentOverriderGroup")
-public class ParentOverrider extends ClassWithJavaDocTag implements InterfaceWithJavaDocTag {
-}
