+++ frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/java5/annotation/ClassWithoutJava5Annotation.java
@@ -1,0 +1,4 @@
+package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.java5.annotation;
+
+public class ClassWithoutJava5Annotation {
+}
