+++ frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/java5/annotation/InheriterFromGrandparent.java
@@ -1,0 +1,5 @@
+package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.java5.annotation;
+
+public class InheriterFromGrandparent extends InheriterFromParent {
+
+}
