+++ frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/java5/annotation/InterfaceWithJavaDocTag.java
@@ -1,4 +1,4 @@
-package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.classtag;
+package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.java5.annotation;
 
 import nl.nn.adapterframework.doc.IbisDoc;
