+++ frankDoc/src/test/java/nl/nn/adapterframework/frankdoc/testtarget/doclet/interfaces/java5/annotation/ParentOverrider.java
@@ -1,0 +1,11 @@
+package nl.nn.adapterframework.frankdoc.testtarget.doclet.interfaces.java5.annotation;
+
+import nl.nn.adapterframework.doc.IbisDoc;
+
+/**
+ * @author martijn
+ *
+ */
+@IbisDoc("ParentOverriderGroup")
+public class ParentOverrider extends ClassWithJava5Annotation implements InterfaceWithJavaDocTag {
+}
